A2/backend/c1/requirements.txt
mysql-connector-python==8.0.23
Flask==1.1.2
flask-cors==3.0.0
